Moved TextCNN to a seperate file
